fix 32 bits build
Former-commit-id: ec69515e037e4f95dac725c92e721edfd008222e

+++ modules/src/render/sdl_audio.cpp
@@ -56,7 +56,7 @@ bool SDLAudio::process(std::shared_ptr<Data> data) {
 
 	{
 		std::lock_guard<std::mutex> lg(m_Mutex);
-		m_Fifo.write(pcmData->data(), pcmData->size());
+		m_Fifo.write(pcmData->data(), (size_t)pcmData->size());
 	}
 
 	return true;
